added support for db pool size from env

+++ src/lib/config/postgres.rs
@@ -5,7 +5,7 @@ pub struct PostgresConfig {
     pub database: String,
     pub user: String,
     pub password: String,
-    pub pool_size: u32,
+    pub pool_size: usize,
     pub connection_timeout: u32,
     pub keepalives_idle: u32,
 }
